update dep, make ui an option
