+++ clashtui/Cargo.toml
@@ -16,18 +16,26 @@ exclude = ["PKGBUILD", "Scoop/*", "Example/*", ".github/*"]
 [badges]
 maintenance = { status = "passively-maintained" }
 
+[features]
+tui = ["ui", "ratatui"]
+
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-ui = { path = "ui" }
+ui = { path = "ui", optional = true }
 api = { path = "api" }
-ratatui = {version = "*", default-features = false, features = ["crossterm"]}
-serde = {version = "*", default-features = false}
+ratatui = { version = "*", default-features = false, features = [
+  "crossterm",
+], optional = true }
+serde = { version = "*", default-features = false }
 argh = "0.1"
 serde_yaml = "0.9"
 serde_json = "1.0"
 log = "0.4"
-log4rs = {version = "1.3", default-features = false, features = ["pattern_encoder", "file_appender"]}
+log4rs = { version = "1.3", default-features = false, features = [
+  "pattern_encoder",
+  "file_appender",
+] }
 enumflags2 = "0.7.9"
 
 [target.'cfg(target_os = "windows")'.dependencies]
@@ -51,7 +59,15 @@ depends = "$auto"
 section = "utility"
 priority = "optional"
 assets = [
-  ['target/release/clashtui', 'usr/bin/clashtui', '755'],
-  ['../README.md', 'usr/share/doc/hust-network-login/README.md', '644'],
+  [
+    'target/release/clashtui',
+    'usr/bin/clashtui',
+    '755',
+  ],
+  [
+    '../README.md',
+    'usr/share/doc/hust-network-login/README.md',
+    '644',
+  ],
 ]
 maintainer-scripts = 'debian/'
