+++ clashtui/src/tui/symbols.rs
@@ -32,11 +32,6 @@ Esc: Close popup
 Q: Quit
 ?: help"#;
 
-pub(crate) const DEFAULT_BASIC_CLASH_CFG_CONTENT: &str = r#"mixed-port: 7890
-mode: rule
-log-level: info
-external-controller: 127.0.0.1:9090"#;
-
 pub(super) const PROFILE: &str = "Profile";
 pub(super) const TEMPALTE: &str = "Template";
 pub(super) const CLASHSRVCTL: &str = "ClashSrvCtl";
