+++ clashtui/src/tui/tabs/mod.rs
@@ -56,55 +56,3 @@ macro_rules! msgpopup_methods {
         }
     };
 }
-#[macro_export]
-macro_rules! define_enum {
-    ($(#[$attr:meta])*
-    $vis:vis $name: ident,
-    [$($variant:ident),*]) => {
-        $(#[$attr])*
-        $vis enum $name {
-            $($variant),*
-        }
-
-        impl From<&str> for $name {
-            fn from(value: &str) -> Self {
-                match value {
-                    $(stringify!($variant) => $name::$variant,)*
-                    _ => panic!("Invalid value for conversion"),
-                }
-            }
-        }
-
-        impl From<$name> for String {
-            fn from(value: $name) -> Self {
-                match value {
-                    $($name::$variant => String::from(stringify!($variant)),)*
-                }
-            }
-        }
-    };
-}
-
-#[cfg(target_os = "linux")]
-define_enum!(
-    pub ClashSrvOp,
-    [
-        StartClashService,
-        StopClashService,
-        SetPermission,
-        SwitchMode
-    ]
-);
-#[cfg(target_os = "windows")]
-define_enum!(
-    pub ClashSrvOp,
-    [
-        StartClashService,
-        StopClashService,
-        SwitchSysProxy,
-        EnableLoopback,
-        InstallSrv,
-        UnInstallSrv,
-        SwitchMode
-    ]
-);
